+++ src/app/(user)/signup/page.tsx
@@ -15,7 +15,7 @@ import { toast } from '@/components/ui/use-toast';
 import { signup } from '@/data/actions/userAction';
 import { zodResolver } from '@hookform/resolvers/zod';
 import Image from 'next/image';
-import { redirect } from 'next/navigation';
+import { useDaumPostcodePopup } from 'react-daum-postcode';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
 
@@ -58,7 +58,11 @@ const FormSchema = z
                 message: '핸드폰 번호는 숫자 형식이어야 합니다.',
             }),
 
-        address: z.string().min(1, { message: '주소를 입력해주세요' }),
+        zoneCode: z.string(),
+
+        roadAddress: z.string(),
+
+        detailAddress: z.string(),
 
         type: z.literal('user'),
     })
@@ -67,6 +71,18 @@ const FormSchema = z
         path: ['passwordCheck'],
         message: '비밀번호가 일치하지 않습니다.',
     });
+
+type Address = {
+    zonecode: string;
+    address: string;
+    addressEnglish: string;
+    addressType: 'R' | 'J';
+    userSelectedType: 'R' | 'J';
+    noSelected: 'Y' | 'N';
+    userLanguageType: 'K' | 'E';
+    roadAddress: string;
+    bname: string;
+};
 
 export default function Signup() {
     // zod resolver - react-hook-form과 zod를 이어주는 다리 역할
@@ -78,12 +94,30 @@ export default function Signup() {
             password: '',
             passwordCheck: '',
             phone: '',
-            address: '',
+            zoneCode: '',
+            roadAddress: '',
+            detailAddress: '',
             type: 'user',
         },
     });
 
-    //& FIXME : toast 모바일 상에서 위치 수정
+    // 주소 API 팝업
+    const open = useDaumPostcodePopup(
+        'https://t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js',
+    );
+
+    // 주소 API 값 선택시 함수 호출
+    const openAddressPopup = () => {
+        open({ onComplete: handleComplete });
+    };
+
+    // 주소 선택시 form에 값 삽입
+    const handleComplete = (data: Address) => {
+        form.setValue('zoneCode', data.zonecode);
+        form.setValue('roadAddress', data.roadAddress);
+    };
+
+    // 회원가입시 formData 전송
     async function onSubmit(formData: z.infer<typeof FormSchema>) {
         // passwordCheck 데이터를 제외를 위한 객체복사
         const { passwordCheck, ...filteredData } = formData;
@@ -215,18 +249,71 @@ export default function Signup() {
                     />
                     <FormField
                         control={form.control}
-                        name='phone'
+                        name='zoneCode'
                         render={({ field }) => (
                             <FormItem className='mb-6'>
-                                <FormLabel htmlFor='phone'>
-                                    휴대폰 번호
+                                <FormLabel htmlFor='zoneCode'>
+                                    우편 번호
+                                </FormLabel>
+                                <FormControl className='flex'>
+                                    <div>
+                                        <Input
+                                            id='zoneCode'
+                                            className='border-0 border-b-[1px] rounded-none p-[5px] border-txt-foreground'
+                                            type='text'
+                                            placeholder='우편번호를 검색하세요'
+                                            {...field}
+                                        />
+                                        <Button
+                                            type='button'
+                                            className='font-notoSansKr right-0 bottom-[0.0625rem]'
+                                            size={'sm'}
+                                            fontSize={'sm'}
+                                            fontWeight={'sm'}
+                                            onClick={openAddressPopup}>
+                                            검색
+                                        </Button>
+                                    </div>
+                                </FormControl>
+                                <FormMessage />
+                            </FormItem>
+                        )}
+                    />
+                    <FormField
+                        control={form.control}
+                        name='roadAddress'
+                        render={({ field }) => (
+                            <FormItem className='mb-6'>
+                                <FormLabel htmlFor='roadAddress'>
+                                    도로명 주소
                                 </FormLabel>
                                 <FormControl className='flex'>
                                     <Input
-                                        id='phone'
+                                        id='roadAddress'
                                         className='border-0 border-b-[1px] rounded-none p-[5px] border-txt-foreground'
                                         type='text'
-                                        placeholder='휴대폰 번호를 입력해주세요'
+                                        placeholder='주소를 입력하세요'
+                                        {...field}
+                                    />
+                                </FormControl>
+                                <FormMessage />
+                            </FormItem>
+                        )}
+                    />
+                    <FormField
+                        control={form.control}
+                        name='detailAddress'
+                        render={({ field }) => (
+                            <FormItem className='mb-6'>
+                                <FormLabel htmlFor='detailAddress'>
+                                    상세 주소
+                                </FormLabel>
+                                <FormControl className='flex'>
+                                    <Input
+                                        id='detailAddress'
+                                        className='border-0 border-b-[1px] rounded-none p-[5px] border-txt-foreground'
+                                        type='text'
+                                        placeholder='상세 주소를 입력하세요'
                                         {...field}
                                     />
                                 </FormControl>
